small fix in get_product_by_name method

+++ services/products.py
@@ -24,7 +24,7 @@ class ProductsService:
             self,
             product_data: Product
     ):
-        prod_try_name = await self.get_product_by_name(product_data.name)
+        prod_try_name = await self.get_product_by_name_helper(product_data.name)
         if prod_try_name:
             prod_try_name.count += product_data.count
             prod_try_name.price = product_data.price
@@ -74,21 +74,27 @@ class ProductsService:
         return prod
 
 
-    async def get_product_by_name(
+    async def get_product_by_name_helper(
             self,
             name: str
     ):
         stmt = select(tables.Products).filter(tables.Products.name == name)
         result = await self.session.execute(stmt)
         product_by_name = result.scalars().first()
-        if product_by_name is None:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail= f"Продукт '{name}' не найден"
-            )
-
         return product_by_name
 
+
+    async def get_by_name(
+            self,
+            name: str
+    ):
+        res = await self.get_product_by_name_helper(name)
+        if res is None:
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail=f"Товар с именем {name} не найден"
+            )
+        return res
     async def get_product_by_id(
             self,
             product_id: int
@@ -121,8 +127,12 @@ class ProductsService:
             self,
             product_name: str,
     ):
-        prod = await self.get_product_by_name(product_name)
-
+        prod = await self.get_product_by_name_helper(product_name)
+        if prod is None:
+            raise HTTPException(
+                        status_code=status.HTTP_404_NOT_FOUND,
+                        detail= f"Продукт '{product_name}' не найден"
+                    )
         await self.session.delete(prod)
         await self.session.commit()
         return JSONResponse(
